timer: fix API.
`expired` method was used for unit-testing, which accepted an optional
argument that is relevant only for unit-testing.

Now `elapsed` is added for unit-testing, which now accepts the optional
argument as required argument and `expired` does not accept any argument
which is cleaner.

Internally both `elapsed` and `expired` call `do_expired`.